Feature/move web request translator (#191)
* Move web request translator

* Fix lint errors

* Fix more lint issues
