+++ CMakeLists.txt
@@ -28,7 +28,6 @@ add_subdirectory(rice-datanode)
 add_subdirectory(native-filesystem)
 add_subdirectory(test)
 add_subdirectory(storage)
-add_subdirectory(web-request)
 add_subdirectory(web-rdfs)
 
 file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/config/hdfs-default.xml
