+++ test/CMakeLists.txt
@@ -25,7 +25,6 @@ include_directories(../rpcserver/include)
 include_directories(../storage/include)
 include_directories(../isal/include)
 include_directories(util/include)
-include_directories(../web-request/include)
 include_directories(../web-rdfs/include)
 
 link_directories(/home/vagrant/rdfs/build/zkwrapper)
@@ -122,4 +121,4 @@ target_link_libraries(WebServerTest ${GTEST})
 target_link_libraries(WebServerTest namenode_protos)
 target_link_libraries(WebServerTest zk_client_nn zk_client_dn rpcserver)
 target_link_libraries(WebServerTest zkwrapper /usr/local/lib/libzookeeper_mt.so)
-target_link_libraries(WebServerTest web-request web_rdfs_server)
+target_link_libraries(WebServerTest web_rdfs_server)
