+++ web-rdfs/CMakeLists.txt
@@ -7,7 +7,6 @@ include_directories(
     ${CMAKE_SOURCE_DIR}/web-rdfs/include
     ${CMAKE_SOURCE_DIR}/zookeeper/include
     ${CMAKE_SOURCE_DIR}/zkwrapper/include
-    ${CMAKE_SOURCE_DIR}/web-request/include
 )
 
 add_library(web_rdfs_server source/web_rdfs_server.cc)
@@ -18,13 +17,17 @@ target_link_libraries(web_rdfs_server zkwrapper)
 target_link_libraries(web_rdfs_server messages )
 target_link_libraries(web_rdfs_server zk_client_dn)
 target_link_libraries(web_rdfs_server zk_client_nn)
-target_link_libraries(web_rdfs_server web-request)
 target_link_libraries(web_rdfs_server ${Boost_LIBRARIES})
+
+add_library(web_request_translator source/web_request_translator.cc)
+target_link_libraries(web_request_translator easylogging++)
+target_link_libraries(web_request_translator protobuf)
 
 add_library(http_handlers source/http_handlers.cc)
 target_link_libraries(http_handlers easylogging++)
-target_link_libraries(http_handlers web-request)
+target_link_libraries(http_handlers web_request_translator)
 
 add_executable(webrdfs source/main.cc)
 target_link_libraries(webrdfs web_rdfs_server)
 target_link_libraries(webrdfs http_handlers)
+
